[packages/ast] add a new package
