+++ apps/desktop/package.json
@@ -8,6 +8,7 @@
     "tauri": "tauri"
   },
   "dependencies": {
+    "@linen/ast": "*",
     "@linen/styles": "*",
     "@linen/ui": "*",
     "@tauri-apps/api": "^1.1.0",
