+++ apps/desktop/src/types.d.ts
@@ -1,2 +1,3 @@
-declare module '@linen.styles';
+declare module '@linen/ast';
+declare module '@linen/styles';
 declare module '@linen/ui';
