+++ apps/web/index.d.ts
@@ -1,5 +1,6 @@
 declare module 'autosize';
 declare module 'lodash.unionby';
 declare module 'lodash.uniq';
+declare module '@linen/ast';
 declare module '@linen/styles';
 declare module '@linen/ui';
