+++ apps/web/package.json
@@ -32,6 +32,7 @@
   },
   "dependencies": {
     "@headlessui/react": "^1.7.3",
+    "@linen/ast": "*",
     "@linen/styles": "*",
     "@linen/ui": "*",
     "@prisma/client": "^3.15.2",
