--- apps/web/utilities/message/walk/index.js
@@ -1,8 +1,0 @@
-const walk = (root, callback) => {
-  callback(root);
-  if (root.children) {
-    root.children.forEach((node) => walk(node, callback));
-  }
-};
-
-export default walk;
