+++ packages/ast/README.md
@@ -1,0 +1,7 @@
+# Abstract Syntax Tree
+
+The package contains parsers, stringifiers and utilities related to various syntaxes used in messages within linen.
+
+The tree structure is easy to work with and simplifies the rendering process.
+
+All parsers should generate same nodes, so we can reuse same rendering process for different message formats.
