+++ packages/ast/index.js
@@ -1,0 +1,11 @@
+const find = require('./find');
+const parse = require('./parse');
+const stringify = require('./stringify');
+const walk = require('./walk');
+
+module.exports = {
+  find,
+  parse,
+  stringify,
+  walk,
+};
