+++ packages/ast/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@linen/ast",
+  "private": true,
+  "version": "1.0.0",
+  "main": "index.js",
+  "files": [
+    "**/*.js"
+  ],
+  "scripts": {
+    "test": "jest"
+  },
+  "devDependencies": {
+    "jest": "^29.3.1"
+  },
+  "dependencies": {
+    "lodash.uniq": "^4.5.0"
+  }
+}
