+++ packages/ast/parse/index.js
@@ -1,0 +1,9 @@
+const discord = require('./discord');
+const linen = require('./linen');
+const slack = require('./slack');
+
+module.exports = {
+  discord,
+  linen,
+  slack,
+};
