+++ packages/ast/parse/linen/expand/index.js
@@ -1,0 +1,108 @@
+const parse = require('../../parse');
+const { explicit, regexp, topline } = require('../../combinators');
+
+const deep = (type, callback) => (match, _, position) => {
+  const [result, content] = match;
+
+  return [
+    {
+      type,
+      children: callback(content),
+      source: result,
+    },
+    position + result.length,
+  ];
+};
+
+const bold = explicit(
+  regexp(
+    /^\*(\S([^*\n]*?|[^*\n]*? `.*?` )\S|\S)\*(?=[\s.,\])}!?\-=]|$)/,
+    deep('bold', (content) => parse(content, matchers))
+  )
+);
+
+const italic = explicit(
+  regexp(
+    /^_(\S([^_\n]*?|[^_\n]*? `.*?` )\S|\S)\_(?=[\s.,\])}!?\-=]|$)/,
+    deep('italic', (content) => parse(content, matchers))
+  )
+);
+
+const strike = explicit(
+  regexp(
+    /^~(\S([^~\n]*?|[^~\n]*? `.*?` )\S|\S)\~(?=[\s.,\])}!?\-=]|$)/,
+    deep('strike', (content) => parse(content, matchers))
+  )
+);
+
+const quote = topline(
+  regexp(/^\> (.*)(\n|$)/, (match, _, position) => {
+    const [result, content] = match;
+
+    const entity = content.match(/^((\> )+)(.*)$/);
+
+    return [
+      {
+        type: 'quote',
+        children: entity
+          ? [
+              {
+                type: 'text',
+                value: entity[1],
+                source: entity[1],
+              },
+              ...parse(entity[3], matchers),
+            ]
+          : parse(content, matchers),
+        source: result,
+      },
+      position + result.length,
+    ];
+  })
+);
+
+const header = topline(
+  regexp(/^\#{1,6} (.*)(\n|$)/, (match, _, position) => {
+    const [result, content] = match;
+
+    const entity = content.match(/^((\#{1, 6} )+)(.*)$/);
+
+    const depth = result.split(' ')[0].length;
+
+    return [
+      {
+        type: 'header',
+        depth,
+        children: entity
+          ? [
+              {
+                type: 'text',
+                value: entity[1],
+                source: entity[1],
+              },
+              ...parse(entity[3], matchers),
+            ]
+          : parse(content, matchers),
+        source: result,
+      },
+      position + result.length,
+    ];
+  })
+);
+
+const matchers = [bold, italic, strike, quote, header];
+
+function expand(tokens) {
+  return tokens.reduce((result, token) => {
+    if (token.type === 'text') {
+      const { value } = token;
+      const nodes = parse(value, matchers);
+      nodes.forEach((node) => result.push(node));
+    } else {
+      result.push(token);
+    }
+    return result;
+  }, []);
+}
+
+module.exports = expand;
