+++ packages/ast/parse/linen/split/index.js
@@ -1,0 +1,58 @@
+const pattern =
+  /(\b(https?):\/\/[-A-Z0-9+&@#\/%?=~_|!:,.;]*[-A-Z0-9+&@#\/%=~_|])/gi;
+
+function split(tokens) {
+  const result = [];
+  for (let index = 0, length = tokens.length; index < length; index++) {
+    const token = tokens[index];
+    if (token.type === 'text') {
+      const { value } = token;
+      const matches = value.match(pattern);
+
+      if (matches) {
+        let indexes = [];
+        let needles = [];
+
+        for (const match of matches) {
+          const start = value.indexOf(match);
+          const end = start + match.length;
+          indexes.push(start);
+          needles.push(start);
+          indexes.push(end);
+        }
+
+        if (indexes[0] !== 0) {
+          indexes.unshift(0);
+        }
+        if (indexes[indexes.length - 1] !== value.length) {
+          indexes.push(value.length);
+        }
+
+        for (let index = 0, length = indexes.length; index < length; index++) {
+          const current = indexes[index];
+          const next = indexes[index + 1];
+          if (typeof current === 'number' && typeof next === 'number') {
+            const text = value.substring(current, next);
+            if (needles.includes(current)) {
+              result.push({
+                type: 'url',
+                url: text,
+                value: text,
+                source: text,
+              });
+            } else {
+              result.push({ type: 'text', value: text, source: text });
+            }
+          }
+        }
+      } else {
+        result.push(token);
+      }
+    } else {
+      result.push(token);
+    }
+  }
+  return result;
+}
+
+module.exports = split;
