+++ packages/ast/walk/index.js
@@ -1,0 +1,8 @@
+const walk = (root, callback) => {
+  callback(root);
+  if (root.children) {
+    root.children.forEach((node) => walk(node, callback));
+  }
+};
+
+module.exports = walk;
